refactor(core)!: move schema-diff types into `db::diff` submodule
Relocate the `*Diff` types from `toasty_core::schema::db::*` into a
dedicated `db::diff` submodule and drop the `Diff` suffix from their
names. The new shape matches the rest of toasty's module conventions:
domain types live in noun-named modules with short names.

  SchemaDiff       -> diff::Schema
  TablesDiff       -> diff::Tables
  TablesDiffItem   -> diff::TablesItem
  ColumnsDiff      -> diff::Columns
  ColumnsDiffItem  -> diff::ColumnsItem
  IndicesDiff      -> diff::Indices
  IndicesDiffItem  -> diff::IndicesItem
  TypesDiff        -> diff::Types
  TypesDiffItem    -> diff::TypesItem
  DiffContext      -> diff::Context
  RenameHints      -> diff::RenameHints (path change only)

The `Driver::generate_migration` trait method now takes
`&diff::Schema<'_>` instead of `&SchemaDiff<'_>`.

BREAKING CHANGE: all `*Diff` types in `toasty_core::schema::db` are
relocated and renamed. Update imports to `db::diff::*` and drop the
`Diff` suffix. External driver implementations must update the
`generate_migration` signature.
