+++ crates/toasty-core/src/schema/db/column.rs
@@ -1,8 +1,7 @@
-use super::{DiffContext, TableId, Type, table};
+use super::{TableId, Type, table};
 use crate::stmt;
 
-use hashbrown::{HashMap, HashSet};
-use std::{fmt, ops::Deref};
+use std::fmt;
 
 /// A column in a database table.
 ///
@@ -110,414 +109,85 @@ impl fmt::Debug for ColumnId {
     }
 }
 
-/// The set of differences between two column lists.
-///
-/// Computed by [`ColumnsDiff::from`] and dereferences to
-/// `Vec<ColumnsDiffItem>` for iteration.
-///
-/// # Examples
-///
-/// ```ignore
-/// use toasty_core::schema::db::{ColumnsDiff, DiffContext, RenameHints, Schema};
-///
-/// let previous = Schema::default();
-/// let next = Schema::default();
-/// let hints = RenameHints::new();
-/// let cx = DiffContext::new(&previous, &next, &hints);
-/// let diff = ColumnsDiff::from(&cx, &[], &[]);
-/// assert!(diff.is_empty());
-/// ```
-pub struct ColumnsDiff<'a> {
-    items: Vec<ColumnsDiffItem<'a>>,
-}
-
-impl<'a> ColumnsDiff<'a> {
-    /// Computes the diff between two column slices.
-    ///
-    /// Uses [`DiffContext`] to resolve rename hints. Columns matched by name
-    /// (or by rename hint) are compared field-by-field; unmatched columns in
-    /// `previous` become drops, and unmatched columns in `next` become adds.
-    pub fn from(cx: &DiffContext<'a>, previous: &'a [Column], next: &'a [Column]) -> Self {
-        fn has_diff(previous: &Column, next: &Column) -> bool {
-            previous.name != next.name
-                || previous.storage_ty != next.storage_ty
-                || previous.nullable != next.nullable
-                || previous.primary_key != next.primary_key
-                || previous.auto_increment != next.auto_increment
-                || previous.versionable != next.versionable
-        }
-
-        let mut items = vec![];
-        let mut add_ids: HashSet<_> = next.iter().map(|next| next.id).collect();
-
-        let next_map =
-            HashMap::<&str, &'a Column>::from_iter(next.iter().map(|to| (to.name.as_str(), to)));
-
-        for previous in previous {
-            let next = if let Some(next_id) = cx.rename_hints().get_column(previous.id) {
-                cx.next().column(next_id)
-            } else if let Some(next) = next_map.get(previous.name.as_str()) {
-                next
-            } else {
-                items.push(ColumnsDiffItem::DropColumn(previous));
-                continue;
-            };
-
-            add_ids.remove(&next.id);
-
-            if has_diff(previous, next) {
-                items.push(ColumnsDiffItem::AlterColumn { previous, next });
-            }
-        }
-
-        for column_id in add_ids {
-            items.push(ColumnsDiffItem::AddColumn(cx.next().column(column_id)));
-        }
-
-        Self { items }
-    }
-
-    /// Returns `true` if there are no column changes.
-    pub const fn is_empty(&self) -> bool {
-        self.items.is_empty()
-    }
-}
-
-impl<'a> Deref for ColumnsDiff<'a> {
-    type Target = Vec<ColumnsDiffItem<'a>>;
-
-    fn deref(&self) -> &Self::Target {
-        &self.items
-    }
-}
-
-/// A single change detected between two column lists.
-pub enum ColumnsDiffItem<'a> {
-    /// A new column was added.
-    AddColumn(&'a Column),
-    /// An existing column was removed.
-    DropColumn(&'a Column),
-    /// A column was modified (name, type, nullability, or other property changed).
-    AlterColumn {
-        /// The column definition before the change.
-        previous: &'a Column,
-        /// The column definition after the change.
-        next: &'a Column,
-    },
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::schema::db::{
-        Column, ColumnId, ColumnsDiff, ColumnsDiffItem, DiffContext, PrimaryKey, RenameHints,
-        Schema, Table, TableId, Type,
-    };
+#[cfg(all(test, feature = "serde"))]
+mod serde_tests {
+    use crate::schema::db::{Column, ColumnId, TableId, Type};
     use crate::stmt;
 
-    fn make_column(
-        table_id: usize,
-        index: usize,
-        name: &str,
-        storage_ty: Type,
-        nullable: bool,
-    ) -> Column {
+    fn base_column() -> Column {
         Column {
             id: ColumnId {
-                table: TableId(table_id),
-                index,
+                table: TableId(0),
+                index: 0,
             },
-            name: name.to_string(),
-            ty: stmt::Type::String, // Simplified for tests
-            storage_ty,
-            nullable,
+            name: "test".to_string(),
+            ty: stmt::Type::String,
+            storage_ty: Type::Text,
+            nullable: false,
             primary_key: false,
             auto_increment: false,
             versionable: false,
         }
     }
 
-    fn make_schema_with_columns(table_id: usize, columns: Vec<Column>) -> Schema {
-        let mut schema = Schema::default();
-        schema.tables.push(Table {
-            id: TableId(table_id),
-            name: "test_table".to_string(),
-            columns,
-            primary_key: PrimaryKey {
-                columns: vec![],
-                index: super::super::IndexId {
-                    table: TableId(table_id),
-                    index: 0,
-                },
-            },
-            indices: vec![],
-        });
-        schema
+    #[test]
+    fn false_booleans_are_omitted() {
+        let toml = toml::to_string(&base_column()).unwrap();
+        assert!(!toml.contains("nullable"), "toml: {toml}");
+        assert!(!toml.contains("primary_key"), "toml: {toml}");
+        assert!(!toml.contains("auto_increment"), "toml: {toml}");
+        assert!(!toml.contains("versionable"), "toml: {toml}");
     }
 
     #[test]
-    fn test_no_diff_same_columns() {
-        let from_cols = vec![
-            make_column(0, 0, "id", Type::Integer(8), false),
-            make_column(0, 1, "name", Type::Text, false),
-        ];
-        let to_cols = vec![
-            make_column(0, 0, "id", Type::Integer(8), false),
-            make_column(0, 1, "name", Type::Text, false),
-        ];
-
-        let from_schema = make_schema_with_columns(0, from_cols.clone());
-        let to_schema = make_schema_with_columns(0, to_cols.clone());
-        let hints = RenameHints::new();
-        let cx = DiffContext::new(&from_schema, &to_schema, &hints);
-
-        let diff = ColumnsDiff::from(&cx, &from_cols, &to_cols);
-        assert!(diff.is_empty());
+    fn nullable_true_is_included() {
+        let col = Column {
+            nullable: true,
+            ..base_column()
+        };
+        let toml = toml::to_string(&col).unwrap();
+        assert!(toml.contains("nullable = true"), "toml: {toml}");
     }
 
     #[test]
-    fn test_add_column() {
-        let from_cols = vec![make_column(0, 0, "id", Type::Integer(8), false)];
-        let to_cols = vec![
-            make_column(0, 0, "id", Type::Integer(8), false),
-            make_column(0, 1, "name", Type::Text, false),
-        ];
-
-        let from_schema = make_schema_with_columns(0, from_cols.clone());
-        let to_schema = make_schema_with_columns(0, to_cols.clone());
-        let hints = RenameHints::new();
-        let cx = DiffContext::new(&from_schema, &to_schema, &hints);
-
-        let diff = ColumnsDiff::from(&cx, &from_cols, &to_cols);
-        assert_eq!(diff.items.len(), 1);
-        assert!(matches!(diff.items[0], ColumnsDiffItem::AddColumn(_)));
-        if let ColumnsDiffItem::AddColumn(col) = diff.items[0] {
-            assert_eq!(col.name, "name");
-        }
+    fn primary_key_true_is_included() {
+        let col = Column {
+            primary_key: true,
+            ..base_column()
+        };
+        let toml = toml::to_string(&col).unwrap();
+        assert!(toml.contains("primary_key = true"), "toml: {toml}");
     }
 
     #[test]
-    fn test_drop_column() {
-        let from_cols = vec![
-            make_column(0, 0, "id", Type::Integer(8), false),
-            make_column(0, 1, "name", Type::Text, false),
-        ];
-        let to_cols = vec![make_column(0, 0, "id", Type::Integer(8), false)];
-
-        let from_schema = make_schema_with_columns(0, from_cols.clone());
-        let to_schema = make_schema_with_columns(0, to_cols.clone());
-        let hints = RenameHints::new();
-        let cx = DiffContext::new(&from_schema, &to_schema, &hints);
-
-        let diff = ColumnsDiff::from(&cx, &from_cols, &to_cols);
-        assert_eq!(diff.items.len(), 1);
-        assert!(matches!(diff.items[0], ColumnsDiffItem::DropColumn(_)));
-        if let ColumnsDiffItem::DropColumn(col) = diff.items[0] {
-            assert_eq!(col.name, "name");
-        }
+    fn auto_increment_true_is_included() {
+        let col = Column {
+            auto_increment: true,
+            ..base_column()
+        };
+        let toml = toml::to_string(&col).unwrap();
+        assert!(toml.contains("auto_increment = true"), "toml: {toml}");
     }
 
     #[test]
-    fn test_alter_column_type() {
-        let from_cols = vec![make_column(0, 0, "id", Type::Integer(8), false)];
-        let to_cols = vec![make_column(0, 0, "id", Type::Text, false)];
-
-        let from_schema = make_schema_with_columns(0, from_cols.clone());
-        let to_schema = make_schema_with_columns(0, to_cols.clone());
-        let hints = RenameHints::new();
-        let cx = DiffContext::new(&from_schema, &to_schema, &hints);
-
-        let diff = ColumnsDiff::from(&cx, &from_cols, &to_cols);
-        assert_eq!(diff.items.len(), 1);
-        assert!(matches!(diff.items[0], ColumnsDiffItem::AlterColumn { .. }));
+    fn missing_bool_fields_deserialize_as_false() {
+        let toml = "name = \"test\"\nty = \"String\"\nstorage_ty = \"Text\"\n\n[id]\ntable = 0\nindex = 0\n";
+        let col: Column = toml::from_str(toml).unwrap();
+        assert!(!col.nullable);
+        assert!(!col.primary_key);
+        assert!(!col.auto_increment);
+        assert!(!col.versionable);
     }
 
     #[test]
-    fn test_alter_column_nullable() {
-        let from_cols = vec![make_column(0, 0, "id", Type::Integer(8), false)];
-        let to_cols = vec![make_column(0, 0, "id", Type::Integer(8), true)];
-
-        let from_schema = make_schema_with_columns(0, from_cols.clone());
-        let to_schema = make_schema_with_columns(0, to_cols.clone());
-        let hints = RenameHints::new();
-        let cx = DiffContext::new(&from_schema, &to_schema, &hints);
-
-        let diff = ColumnsDiff::from(&cx, &from_cols, &to_cols);
-        assert_eq!(diff.items.len(), 1);
-        assert!(matches!(diff.items[0], ColumnsDiffItem::AlterColumn { .. }));
-    }
-
-    #[test]
-    fn test_rename_column_with_hint() {
-        // Column renamed from "old_name" to "new_name"
-        let from_cols = vec![make_column(0, 0, "old_name", Type::Text, false)];
-        let to_cols = vec![make_column(0, 0, "new_name", Type::Text, false)];
-
-        let from_schema = make_schema_with_columns(0, from_cols.clone());
-        let to_schema = make_schema_with_columns(0, to_cols.clone());
-
-        let mut hints = RenameHints::new();
-        hints.add_column_hint(
-            ColumnId {
-                table: TableId(0),
-                index: 0,
-            },
-            ColumnId {
-                table: TableId(0),
-                index: 0,
-            },
-        );
-        let cx = DiffContext::new(&from_schema, &to_schema, &hints);
-
-        let diff = ColumnsDiff::from(&cx, &from_cols, &to_cols);
-        assert_eq!(diff.items.len(), 1);
-        assert!(matches!(diff.items[0], ColumnsDiffItem::AlterColumn { .. }));
-        if let ColumnsDiffItem::AlterColumn { previous, next } = diff.items[0] {
-            assert_eq!(previous.name, "old_name");
-            assert_eq!(next.name, "new_name");
-        }
-    }
-
-    #[test]
-    fn test_rename_column_without_hint_is_drop_and_add() {
-        // Column renamed from "old_name" to "new_name", but no hint provided
-        // Should be treated as drop + add
-        let from_cols = vec![make_column(0, 0, "old_name", Type::Text, false)];
-        let to_cols = vec![make_column(0, 0, "new_name", Type::Text, false)];
-
-        let from_schema = make_schema_with_columns(0, from_cols.clone());
-        let to_schema = make_schema_with_columns(0, to_cols.clone());
-        let hints = RenameHints::new();
-        let cx = DiffContext::new(&from_schema, &to_schema, &hints);
-
-        let diff = ColumnsDiff::from(&cx, &from_cols, &to_cols);
-        assert_eq!(diff.items.len(), 2);
-
-        let has_drop = diff
-            .items
-            .iter()
-            .any(|item| matches!(item, ColumnsDiffItem::DropColumn(_)));
-        let has_add = diff
-            .items
-            .iter()
-            .any(|item| matches!(item, ColumnsDiffItem::AddColumn(_)));
-        assert!(has_drop);
-        assert!(has_add);
-    }
-
-    #[cfg(feature = "serde")]
-    mod serde_tests {
-        use crate::schema::db::{Column, ColumnId, TableId, Type};
-        use crate::stmt;
-
-        fn base_column() -> Column {
-            Column {
-                id: ColumnId {
-                    table: TableId(0),
-                    index: 0,
-                },
-                name: "test".to_string(),
-                ty: stmt::Type::String,
-                storage_ty: Type::Text,
-                nullable: false,
-                primary_key: false,
-                auto_increment: false,
-                versionable: false,
-            }
-        }
-
-        #[test]
-        fn false_booleans_are_omitted() {
-            let toml = toml::to_string(&base_column()).unwrap();
-            assert!(!toml.contains("nullable"), "toml: {toml}");
-            assert!(!toml.contains("primary_key"), "toml: {toml}");
-            assert!(!toml.contains("auto_increment"), "toml: {toml}");
-            assert!(!toml.contains("versionable"), "toml: {toml}");
-        }
-
-        #[test]
-        fn nullable_true_is_included() {
-            let col = Column {
-                nullable: true,
-                ..base_column()
-            };
-            let toml = toml::to_string(&col).unwrap();
-            assert!(toml.contains("nullable = true"), "toml: {toml}");
-        }
-
-        #[test]
-        fn primary_key_true_is_included() {
-            let col = Column {
-                primary_key: true,
-                ..base_column()
-            };
-            let toml = toml::to_string(&col).unwrap();
-            assert!(toml.contains("primary_key = true"), "toml: {toml}");
-        }
-
-        #[test]
-        fn auto_increment_true_is_included() {
-            let col = Column {
-                auto_increment: true,
-                ..base_column()
-            };
-            let toml = toml::to_string(&col).unwrap();
-            assert!(toml.contains("auto_increment = true"), "toml: {toml}");
-        }
-
-        #[test]
-        fn missing_bool_fields_deserialize_as_false() {
-            let toml = "name = \"test\"\nty = \"String\"\nstorage_ty = \"Text\"\n\n[id]\ntable = 0\nindex = 0\n";
-            let col: Column = toml::from_str(toml).unwrap();
-            assert!(!col.nullable);
-            assert!(!col.primary_key);
-            assert!(!col.auto_increment);
-            assert!(!col.versionable);
-        }
-
-        #[test]
-        fn round_trip_all_true() {
-            let original = Column {
-                nullable: true,
-                primary_key: true,
-                auto_increment: true,
-                ..base_column()
-            };
-            let decoded: Column = toml::from_str(&toml::to_string(&original).unwrap()).unwrap();
-            assert_eq!(original, decoded);
-        }
-    }
-
-    #[test]
-    fn test_multiple_operations() {
-        let from_cols = vec![
-            make_column(0, 0, "id", Type::Integer(8), false),
-            make_column(0, 1, "old_name", Type::Text, false),
-            make_column(0, 2, "to_drop", Type::Text, false),
-        ];
-        let to_cols = vec![
-            make_column(0, 0, "id", Type::Text, false), // type changed
-            make_column(0, 1, "new_name", Type::Text, false), // renamed
-            make_column(0, 2, "added", Type::Integer(8), false), // new column
-        ];
-
-        let from_schema = make_schema_with_columns(0, from_cols.clone());
-        let to_schema = make_schema_with_columns(0, to_cols.clone());
-
-        let mut hints = RenameHints::new();
-        hints.add_column_hint(
-            ColumnId {
-                table: TableId(0),
-                index: 1,
-            },
-            ColumnId {
-                table: TableId(0),
-                index: 1,
-            },
-        );
-        let cx = DiffContext::new(&from_schema, &to_schema, &hints);
-
-        let diff = ColumnsDiff::from(&cx, &from_cols, &to_cols);
-        // Should have: 1 alter (id type changed), 1 alter (renamed), 1 drop (to_drop), 1 add (added)
-        assert_eq!(diff.items.len(), 4);
+    fn round_trip_all_true() {
+        let original = Column {
+            nullable: true,
+            primary_key: true,
+            auto_increment: true,
+            ..base_column()
+        };
+        let decoded: Column = toml::from_str(&toml::to_string(&original).unwrap()).unwrap();
+        assert_eq!(original, decoded);
     }
 }
