+++ crates/toasty-core/src/schema/db/diff.rs
@@ -1,6 +1,35 @@
+//! Schema-diff types.
+//!
+//! Compares two [`Schema`](super::Schema) versions and produces structured
+//! changes consumed by drivers to generate migrations.
+//!
+//! # Examples
+//!
+//! ```ignore
+//! use toasty_core::schema::db::{Schema, diff};
+//!
+//! let previous = Schema::default();
+//! let next = Schema::default();
+//! let hints = diff::RenameHints::new();
+//! let d = diff::Schema::from(&previous, &next, &hints);
+//! assert!(d.is_empty());
+//! ```
+
+mod columns;
+mod indices;
+mod schema;
+mod tables;
+mod types;
+
+pub use columns::{Columns, ColumnsItem};
+pub use indices::{Indices, IndicesItem};
+pub use schema::Schema;
+pub use tables::{Tables, TablesItem};
+pub use types::{Types, TypesItem};
+
 use hashbrown::HashMap;
 
-use crate::schema::db::{ColumnId, IndexId, Schema, TableId};
+use crate::schema::db::{ColumnId, IndexId, Schema as DbSchema, TableId};
 
 /// Hints that tell the diff algorithm which schema items were renamed.
 ///
@@ -11,9 +40,9 @@ use crate::schema::db::{ColumnId, IndexId, Schema, TableId};
 /// # Examples
 ///
 /// ```ignore
-/// use toasty_core::schema::db::{RenameHints, TableId};
+/// use toasty_core::schema::db::{TableId, diff};
 ///
-/// let mut hints = RenameHints::new();
+/// let mut hints = diff::RenameHints::new();
 /// hints.add_table_hint(TableId(0), TableId(1));
 /// assert_eq!(hints.get_table(TableId(0)), Some(TableId(1)));
 /// assert_eq!(hints.get_table(TableId(2)), None);
@@ -64,31 +93,31 @@ impl RenameHints {
 
 /// Shared context passed to all diff computations.
 ///
-/// Holds references to both the previous and next [`Schema`] versions and
-/// the [`RenameHints`] that guide rename detection.
+/// Holds references to both the previous and next [`Schema`](super::Schema)
+/// versions and the [`RenameHints`] that guide rename detection.
 ///
 /// # Examples
 ///
 /// ```ignore
-/// use toasty_core::schema::db::{DiffContext, RenameHints, Schema};
+/// use toasty_core::schema::db::{Schema, diff};
 ///
 /// let previous = Schema::default();
 /// let next = Schema::default();
-/// let hints = RenameHints::new();
-/// let cx = DiffContext::new(&previous, &next, &hints);
+/// let hints = diff::RenameHints::new();
+/// let cx = diff::Context::new(&previous, &next, &hints);
 /// assert!(cx.previous().tables.is_empty());
 /// ```
-pub struct DiffContext<'a> {
-    previous: &'a Schema,
-    next: &'a Schema,
+pub struct Context<'a> {
+    previous: &'a DbSchema,
+    next: &'a DbSchema,
 
     rename_hints: &'a RenameHints,
 }
 
-impl<'a> DiffContext<'a> {
+impl<'a> Context<'a> {
     /// Creates a new diff context from the previous schema, the next schema,
     /// and the rename hints that map old IDs to new IDs.
-    pub fn new(previous: &'a Schema, next: &'a Schema, rename_hints: &'a RenameHints) -> Self {
+    pub fn new(previous: &'a DbSchema, next: &'a DbSchema, rename_hints: &'a RenameHints) -> Self {
         Self {
             previous,
             next,
@@ -102,12 +131,12 @@ impl<'a> DiffContext<'a> {
     }
 
     /// Returns the schema before the change.
-    pub fn previous(&self) -> &'a Schema {
+    pub fn previous(&self) -> &'a DbSchema {
         self.previous
     }
 
     /// Returns the schema after the change.
-    pub fn next(&self) -> &'a Schema {
+    pub fn next(&self) -> &'a DbSchema {
         self.next
     }
 }
