+++ crates/toasty-sql/tests/migration_alter_column.rs
@@ -1,9 +1,6 @@
 use toasty_core::{
     driver::Capability,
-    schema::db::{
-        Column, ColumnId, IndexId, PrimaryKey, RenameHints, Schema, SchemaDiff, Table, TableId,
-        Type,
-    },
+    schema::db::{Column, ColumnId, IndexId, PrimaryKey, Schema, Table, TableId, Type, diff},
     stmt as core_stmt,
 };
 use toasty_sql::{Serializer, migration::MigrationStatement};
@@ -89,7 +86,7 @@ fn alter_column_rename_postgresql() {
         )],
     };
 
-    let mut hints = RenameHints::new();
+    let mut hints = diff::RenameHints::new();
     hints.add_column_hint(
         ColumnId {
             table: TableId(0),
@@ -101,7 +98,7 @@ fn alter_column_rename_postgresql() {
         },
     );
 
-    let diff = SchemaDiff::from(&from, &to, &hints);
+    let diff = diff::Schema::from(&from, &to, &hints);
     let stmts = MigrationStatement::from_diff(&diff, &Capability::POSTGRESQL);
     let sql = serialize_migration(&stmts, "postgresql");
 
@@ -135,7 +132,7 @@ fn alter_column_rename_with_table_rename_postgresql() {
         )],
     };
 
-    let mut hints = RenameHints::new();
+    let mut hints = diff::RenameHints::new();
     hints.add_table_hint(TableId(0), TableId(0));
     hints.add_column_hint(
         ColumnId {
@@ -148,7 +145,7 @@ fn alter_column_rename_with_table_rename_postgresql() {
         },
     );
 
-    let diff = SchemaDiff::from(&from, &to, &hints);
+    let diff = diff::Schema::from(&from, &to, &hints);
     let stmts = MigrationStatement::from_diff(&diff, &Capability::POSTGRESQL);
     let sql = serialize_migration(&stmts, "postgresql");
 
@@ -184,8 +181,8 @@ fn alter_column_set_not_null_postgresql() {
         )],
     };
 
-    let hints = RenameHints::new();
-    let diff = SchemaDiff::from(&from, &to, &hints);
+    let hints = diff::RenameHints::new();
+    let diff = diff::Schema::from(&from, &to, &hints);
     let stmts = MigrationStatement::from_diff(&diff, &Capability::POSTGRESQL);
     let sql = serialize_migration(&stmts, "postgresql");
 
@@ -219,8 +216,8 @@ fn alter_column_drop_not_null_postgresql() {
         )],
     };
 
-    let hints = RenameHints::new();
-    let diff = SchemaDiff::from(&from, &to, &hints);
+    let hints = diff::RenameHints::new();
+    let diff = diff::Schema::from(&from, &to, &hints);
     let stmts = MigrationStatement::from_diff(&diff, &Capability::POSTGRESQL);
     let sql = serialize_migration(&stmts, "postgresql");
 
@@ -253,8 +250,8 @@ fn alter_column_change_type_postgresql() {
         )],
     };
 
-    let hints = RenameHints::new();
-    let diff = SchemaDiff::from(&from, &to, &hints);
+    let hints = diff::RenameHints::new();
+    let diff = diff::Schema::from(&from, &to, &hints);
     let stmts = MigrationStatement::from_diff(&diff, &Capability::POSTGRESQL);
     let sql = serialize_migration(&stmts, "postgresql");
 
@@ -288,8 +285,8 @@ fn alter_column_multiple_changes_postgresql() {
         )],
     };
 
-    let hints = RenameHints::new();
-    let diff = SchemaDiff::from(&from, &to, &hints);
+    let hints = diff::RenameHints::new();
+    let diff = diff::Schema::from(&from, &to, &hints);
     let stmts = MigrationStatement::from_diff(&diff, &Capability::POSTGRESQL);
     let sql = serialize_migration(&stmts, "postgresql");
 
@@ -329,7 +326,7 @@ fn alter_column_rename_and_change_type_postgresql() {
         )],
     };
 
-    let mut hints = RenameHints::new();
+    let mut hints = diff::RenameHints::new();
     hints.add_column_hint(
         ColumnId {
             table: TableId(0),
@@ -341,7 +338,7 @@ fn alter_column_rename_and_change_type_postgresql() {
         },
     );
 
-    let diff = SchemaDiff::from(&from, &to, &hints);
+    let diff = diff::Schema::from(&from, &to, &hints);
     let stmts = MigrationStatement::from_diff(&diff, &Capability::POSTGRESQL);
     let sql = serialize_migration(&stmts, "postgresql");
 
@@ -382,7 +379,7 @@ fn alter_column_rename_and_drop_not_null_postgresql() {
         )],
     };
 
-    let mut hints = RenameHints::new();
+    let mut hints = diff::RenameHints::new();
     hints.add_column_hint(
         ColumnId {
             table: TableId(0),
@@ -394,7 +391,7 @@ fn alter_column_rename_and_drop_not_null_postgresql() {
         },
     );
 
-    let diff = SchemaDiff::from(&from, &to, &hints);
+    let diff = diff::Schema::from(&from, &to, &hints);
     let stmts = MigrationStatement::from_diff(&diff, &Capability::POSTGRESQL);
     let sql = serialize_migration(&stmts, "postgresql");
 
@@ -430,10 +427,10 @@ fn alter_column_multiple_changes_with_table_rename_postgresql() {
         )],
     };
 
-    let mut hints = RenameHints::new();
+    let mut hints = diff::RenameHints::new();
     hints.add_table_hint(TableId(0), TableId(0));
 
-    let diff = SchemaDiff::from(&from, &to, &hints);
+    let diff = diff::Schema::from(&from, &to, &hints);
     let stmts = MigrationStatement::from_diff(&diff, &Capability::POSTGRESQL);
     let sql = serialize_migration(&stmts, "postgresql");
 
@@ -475,7 +472,7 @@ fn alter_column_rename_mysql() {
         )],
     };
 
-    let mut hints = RenameHints::new();
+    let mut hints = diff::RenameHints::new();
     hints.add_column_hint(
         ColumnId {
             table: TableId(0),
@@ -487,7 +484,7 @@ fn alter_column_rename_mysql() {
         },
     );
 
-    let diff = SchemaDiff::from(&from, &to, &hints);
+    let diff = diff::Schema::from(&from, &to, &hints);
     let stmts = MigrationStatement::from_diff(&diff, &Capability::MYSQL);
     let sql = serialize_migration(&stmts, "mysql");
 
@@ -520,8 +517,8 @@ fn alter_column_multiple_changes_mysql() {
         )],
     };
 
-    let hints = RenameHints::new();
-    let diff = SchemaDiff::from(&from, &to, &hints);
+    let hints = diff::RenameHints::new();
+    let diff = diff::Schema::from(&from, &to, &hints);
     let stmts = MigrationStatement::from_diff(&diff, &Capability::MYSQL);
     let sql = serialize_migration(&stmts, "mysql");
 
@@ -554,10 +551,10 @@ fn alter_column_multiple_changes_with_table_rename_mysql() {
         )],
     };
 
-    let mut hints = RenameHints::new();
+    let mut hints = diff::RenameHints::new();
     hints.add_table_hint(TableId(0), TableId(0));
 
-    let diff = SchemaDiff::from(&from, &to, &hints);
+    let diff = diff::Schema::from(&from, &to, &hints);
     let stmts = MigrationStatement::from_diff(&diff, &Capability::MYSQL);
     let sql = serialize_migration(&stmts, "mysql");
 
@@ -598,7 +595,7 @@ fn alter_column_rename_only_sqlite() {
         )],
     };
 
-    let mut hints = RenameHints::new();
+    let mut hints = diff::RenameHints::new();
     hints.add_column_hint(
         ColumnId {
             table: TableId(0),
@@ -610,7 +607,7 @@ fn alter_column_rename_only_sqlite() {
         },
     );
 
-    let diff = SchemaDiff::from(&from, &to, &hints);
+    let diff = diff::Schema::from(&from, &to, &hints);
     let stmts = MigrationStatement::from_diff(&diff, &Capability::SQLITE);
     let sql = serialize_migration(&stmts, "sqlite");
 
@@ -648,8 +645,8 @@ fn alter_column_change_nullability_sqlite() {
         )],
     };
 
-    let hints = RenameHints::new();
-    let diff = SchemaDiff::from(&from, &to, &hints);
+    let hints = diff::RenameHints::new();
+    let diff = diff::Schema::from(&from, &to, &hints);
     let stmts = MigrationStatement::from_diff(&diff, &Capability::SQLITE);
     let sql = serialize_migration(&stmts, "sqlite");
 
@@ -690,8 +687,8 @@ fn alter_column_change_type_sqlite() {
         )],
     };
 
-    let hints = RenameHints::new();
-    let diff = SchemaDiff::from(&from, &to, &hints);
+    let hints = diff::RenameHints::new();
+    let diff = diff::Schema::from(&from, &to, &hints);
     let stmts = MigrationStatement::from_diff(&diff, &Capability::SQLITE);
     let sql = serialize_migration(&stmts, "sqlite");
 
@@ -732,10 +729,10 @@ fn alter_column_change_nullability_with_table_rename_sqlite() {
         )],
     };
 
-    let mut hints = RenameHints::new();
+    let mut hints = diff::RenameHints::new();
     hints.add_table_hint(TableId(0), TableId(0));
 
-    let diff = SchemaDiff::from(&from, &to, &hints);
+    let diff = diff::Schema::from(&from, &to, &hints);
     let stmts = MigrationStatement::from_diff(&diff, &Capability::SQLITE);
     let sql = serialize_migration(&stmts, "sqlite");
 
@@ -780,7 +777,7 @@ fn alter_column_rename_and_change_type_sqlite() {
         )],
     };
 
-    let mut hints = RenameHints::new();
+    let mut hints = diff::RenameHints::new();
     hints.add_column_hint(
         ColumnId {
             table: TableId(0),
@@ -792,7 +789,7 @@ fn alter_column_rename_and_change_type_sqlite() {
         },
     );
 
-    let diff = SchemaDiff::from(&from, &to, &hints);
+    let diff = diff::Schema::from(&from, &to, &hints);
     let stmts = MigrationStatement::from_diff(&diff, &Capability::SQLITE);
     let sql = serialize_migration(&stmts, "sqlite");
