+++ crates/toasty-sql/tests/migration_alter_table.rs
@@ -1,9 +1,6 @@
 use toasty_core::{
     driver::Capability,
-    schema::db::{
-        Column, ColumnId, IndexId, PrimaryKey, RenameHints, Schema, SchemaDiff, Table, TableId,
-        Type,
-    },
+    schema::db::{Column, ColumnId, IndexId, PrimaryKey, Schema, Table, TableId, Type, diff},
     stmt as core_stmt,
 };
 use toasty_sql::{Serializer, migration::MigrationStatement};
@@ -84,10 +81,10 @@ fn rename_table_sqlite() {
         )],
     };
 
-    let mut hints = RenameHints::new();
+    let mut hints = diff::RenameHints::new();
     hints.add_table_hint(TableId(0), TableId(0));
 
-    let diff = SchemaDiff::from(&from, &to, &hints);
+    let diff = diff::Schema::from(&from, &to, &hints);
     let stmts = MigrationStatement::from_diff(&diff, &Capability::SQLITE);
     let sql = serialize_migration(&stmts, "sqlite");
 
@@ -118,10 +115,10 @@ fn rename_table_postgresql() {
         )],
     };
 
-    let mut hints = RenameHints::new();
+    let mut hints = diff::RenameHints::new();
     hints.add_table_hint(TableId(0), TableId(0));
 
-    let diff = SchemaDiff::from(&from, &to, &hints);
+    let diff = diff::Schema::from(&from, &to, &hints);
     let stmts = MigrationStatement::from_diff(&diff, &Capability::POSTGRESQL);
     let sql = serialize_migration(&stmts, "postgresql");
 
@@ -149,10 +146,10 @@ fn rename_table_and_add_column() {
         )],
     };
 
-    let mut hints = RenameHints::new();
+    let mut hints = diff::RenameHints::new();
     hints.add_table_hint(TableId(0), TableId(0));
 
-    let diff = SchemaDiff::from(&from, &to, &hints);
+    let diff = diff::Schema::from(&from, &to, &hints);
     let stmts = MigrationStatement::from_diff(&diff, &Capability::SQLITE);
     let sql = serialize_migration(&stmts, "sqlite");
 
@@ -187,8 +184,8 @@ fn rename_without_hint_is_drop_and_create() {
         )],
     };
 
-    let hints = RenameHints::new();
-    let diff = SchemaDiff::from(&from, &to, &hints);
+    let hints = diff::RenameHints::new();
+    let diff = diff::Schema::from(&from, &to, &hints);
     let stmts = MigrationStatement::from_diff(&diff, &Capability::SQLITE);
     let sql = serialize_migration(&stmts, "sqlite");
