+++ crates/toasty-sql/tests/migration_drop_column.rs
@@ -1,9 +1,6 @@
 use toasty_core::{
     driver::Capability,
-    schema::db::{
-        Column, ColumnId, IndexId, PrimaryKey, RenameHints, Schema, SchemaDiff, Table, TableId,
-        Type,
-    },
+    schema::db::{Column, ColumnId, IndexId, PrimaryKey, Schema, Table, TableId, Type, diff},
     stmt as core_stmt,
 };
 use toasty_sql::{Serializer, migration::MigrationStatement};
@@ -81,8 +78,8 @@ fn drop_column_sqlite() {
         )],
     };
 
-    let hints = RenameHints::new();
-    let diff = SchemaDiff::from(&from, &to, &hints);
+    let hints = diff::RenameHints::new();
+    let diff = diff::Schema::from(&from, &to, &hints);
     let stmts = MigrationStatement::from_diff(&diff, &Capability::SQLITE);
     let sql = serialize_migration(&stmts, "sqlite");
 
@@ -110,8 +107,8 @@ fn drop_column_postgresql() {
         )],
     };
 
-    let hints = RenameHints::new();
-    let diff = SchemaDiff::from(&from, &to, &hints);
+    let hints = diff::RenameHints::new();
+    let diff = diff::Schema::from(&from, &to, &hints);
     let stmts = MigrationStatement::from_diff(&diff, &Capability::POSTGRESQL);
     let sql = serialize_migration(&stmts, "postgresql");
 
@@ -140,8 +137,8 @@ fn drop_multiple_columns() {
         )],
     };
 
-    let hints = RenameHints::new();
-    let diff = SchemaDiff::from(&from, &to, &hints);
+    let hints = diff::RenameHints::new();
+    let diff = diff::Schema::from(&from, &to, &hints);
     let stmts = MigrationStatement::from_diff(&diff, &Capability::SQLITE);
     let sql = serialize_migration(&stmts, "sqlite");
